+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Change Log
 ----------
 
+1.0.2
+
+- Rename config key names.
+
+1.0.1
+
+- Support parse xmind to xml file type.
+
 1.0.0
 
 - Support parse xmind to dict data type with Python.